+++ packages/odata/index.ts
@@ -1,9 +1,8 @@
 export * from "./batch";
 export * from "./caching";
-export * from "./errors";
 export * from "./extenders";
-export * from "./invokable";
-export * from "./operation-binder";
+export * from "./invokable-binder";
+export * from "./pipeline-binder";
 export * from "./parsers";
 export * from "./pipeline";
 export * from "./queryable";
